Fixed npm installation and moved to types folder.
