+++ .github/workflows/windows.yml
@@ -116,6 +116,8 @@ jobs:
       -
         name: Install Cypress
         run: cd etherpad && cd src && npm install cypress --legacy-peer-deps
+      - name: Install npm@6
+        run: npm install npm@6 -g
       -
         name: Run Etherpad
         run: |
